requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.6
django-allauth==0.58.2
django-cors-headers==4.3.0
django-environ==0.11.2
django-filter==23.4
djangorestframework==3.14.0
gunicorn==21.2.0
haversine==2.8.0
idna==3.6
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.1
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing-extensions==4.8.0
urllib3==2.1.0
whitenoise==6.6.0
